feat(ehr): share command palette search button across headers
Extract the search-styled launcher into CommandPaletteSearchButton so the
main nav, in-person progress note header, and telemed appointment header
all use one implementation. Shortcut hint is platform-aware (⌘K on Mac,
Ctrl+K elsewhere). In-person header places it inline with the close
button, at the same row as the provider dropdown.
